Agenda : remove events
an event that is marked deleted but with a startDate in the past is removed from the service plan collection and no message is sent to hhr-home. ( the event of the service is considered consumed ).

for an event marked as deleted with a start date greater or equal to the current date,  we send a delete message to the hhr-home (event canceled and not yet consumed).