Fix #304 : Update of the e-mail sent to participants
it's now possible to send a custom email to beneficiaries.

to do that, a mailTpl property should be define in the config.json with the path of the email template ( typically in the /config directory ) without the extension :

for example : /config/mailtobeneficiary

the two files ( /config/mailtobeneficiary.htm & /config/mailtobeneficiary.txt ) should be present otherwise the app will use the default template.